Cleaning up unnecessary imports across the entire project. Storing and representing Item timestamp as a more natural long instead of string. Moved the 'getTimeStampString' method from 'ItemDataSource' to 'Item' to better support the change of the timeStamp data type from String to Long.

+++ app/src/main/java/com/example/skybox/groceree/SelectionAdapter.java
@@ -66,7 +66,7 @@ public class SelectionAdapter extends ArrayAdapter<Item> {
         TextView tvTimeStamp = ( TextView ) convertView.findViewById( R.id.textViewTimeStamp );
 
         tvItem.setText( item.toString() );
-        tvTimeStamp.setText( Integer.toString( item.getTimeStamp() ) );
+        tvTimeStamp.setText( Long.toString( item.getTimeStamp() ) );
 
         if( mSelectedItems.get( position ) != false ) {
             convertView.setBackgroundColor( context.getResources().getColor( android.R.color.holo_blue_light ) );
